Don't block on the send log.
Use the sqlite online backup api to take snapshots of the send log and
then apply a new send long state over the old one at the end. This has one
downside, which is last put wins when it comes to send logging.
In general users should use explicit send logs if they care,
so this does not seem like a bad tradeoff.