Use an address cache for put and refactor send log.
Use of an address cache saves lot of network io if we disable the send
log. We also refactored the send log to reduce the number of sqlite3
statements by one on each send log cache hit.

For my personal backups this address cache reduces network IO by
a quarter.