Mitigate timestamp spoofing.
Previously timestamps were encrypted by the client, but this meant the
server had no way to verify them. Now both the server and client must
independently verify a metadata timestamp is valid before accepting.

This change warrants a new metadata version as it could be used during
an attack to spoof timestamps in a way that could confuse someone trying
to recover a system, even if they knew the date of the breach.

We also took this opportunity to change complex string timestamps to a
simpler unix millisecond timestamp.