OpenBSD only: use pledge(2) and unveil(2) (#103)
Ideally unveil(2) would be used once in `main.rs` to unveil repodir from
a restricted file system view. This does not work for `bupstash init` as
repodir does not exist yet. For this operation additional calls to
unveil(2) is used to unveil repodir.bupstash-repo-init-tmp.
Additionally an unveil(2) is required in `fsutil.rs`.

Use of unveil(2) and pledge(2) is limited to `bupstash serve`.