Fix ordering problem in index writing, make indexing parallel.
This change refactors the filesystem indexer such that it walks the disk
using a stack. This change was needed as the old logic produced indexes
that did not work properly when combined with diff.

Incidental to the change, the indexer was made more parallel as this
was trivial during the required refactor.