Migrate from glob -> globset
The initial reason was because of $features, but actually globset
should also give us better performance when there are many exclusions.

Note that we rely on custom patches for globset until they are merged
upstream. Since forked dependencies may not be used for crates.io crates,
we instead use type wrappers to provide the functionality locally.